User Editing: Created Output actor and moved over relevant code. Using event emitter for IPC between pdfUploader and Ouput.
