+++ src/eventEmitter.js
@@ -1,0 +1,5 @@
+import EventEmitter from 'eventemitter3';
+
+const eventEmitter = new EventEmitter();
+
+export default eventEmitter;
